+++ lib/aletheia_core_model.dart
@@ -21,10 +21,14 @@ export 'model/dictionary/dictionary.word.dart';
 //
 export 'model/lexico/lexico.dictionary.dart';
 export 'model/lexico/lexico.version.dart';
-//
+//NLP
 export 'model/nlp/analize.grammar.dart';
 export 'model/nlp/nlp.verse.dart';
 export 'model/nlp/type.word.part.of.speech.dart';
+export 'model/nlp/type.word.dependency.edge.dart';
+export 'model/nlp/analize.text.dart';
+export 'model/nlp/type.entity.dart';
+
 //
 export 'model/user/config/user.developer.dart';
 export 'model/user/config/user.google.service.account.credencial.dart';